server/requirements.txt
diffusers
transformers
accelerate
peft
imageio
fastapi[all]
pillow
httpx
starlette
